+++ src/advanced/widgets/cart-sidebar/index.ts
@@ -1,0 +1,1 @@
+export { CartSidebar } from './ui/CartSidebar';
