+++ src/advanced/widgets/shopping-product-list/ui/ProductList.tsx
@@ -1,8 +1,8 @@
-import type { CartItem } from '../../../types';
-import type { ProductWithUI } from '../../entities/product';
-import { filterProducts, getRemainingStock } from '../../entities/product';
-import { Icon } from '../../shared/icon';
-import { convertRateToPercentage, formatKRWPrice } from '../../shared/lib';
+import type { CartItem } from '../../../../types';
+import type { ProductWithUI } from '../../../entities/product';
+import { filterProducts, getRemainingStock } from '../../../entities/product';
+import { Icon } from '../../../shared/icon';
+import { convertRateToPercentage, formatKRWPrice } from '../../../shared/lib';
 
 interface ProductListProps {
   products: ProductWithUI[];
